requirements.txt
pandas
google-generativeai
tenacity
tkinter
python-dotenv
